+++ src/main/java/com/ihren/processor/processor/TransactionProcessor.java
@@ -1,6 +1,6 @@
 package com.ihren.processor.processor;
 
-import com.ihren.processor.dto.InputTransaction;
+import com.ihren.processor.model.input.InputTransaction;
 import com.ihren.processor.exception.handler.ExceptionHandler;
 import com.ihren.processor.mapper.TransactionMapper;
 import com.ihren.processor.model.Transaction;
