+++ src/test/java/com/ihren/processor/mapper/ItemMapperImplTest.java
@@ -1,6 +1,6 @@
 package com.ihren.processor.mapper;
 
-import com.ihren.processor.dto.InputItem;
+import com.ihren.processor.model.input.InputItem;
 import com.ihren.processor.mapper.exception.MappingException;
 import com.ihren.processor.model.Item;
 import org.junit.jupiter.api.Test;
